%%  WARNING: This file was automatically generated from the associated 
%%  wav_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the wav_data.mid file in the project MID directory. Once the 
%%  wav_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the D.table.critical function

\name{D.table.critical}
\alias{D.table.critical}
\title{Critical D-Statistic Table}
\concept{statistical tables}
\description{The critical D-statistics define the distribution of D for a
zero mean Gaussian white noise process. Comparing a D-statistic
to the corresponding critical values provides a means of quantitatively
rejecting or accepting a linear cumulative energy hypothesis. The table
is generated for an ensemble of distribution probabilities and sample sizes.}

\seealso{
\code{\link{D.table}}, \code{\link{D.statistic}}.}

\examples{
D.table.critical[1:10,]
}
\keyword{distribution}

