%%  WARNING: This file was automatically generated from the associated 
%%  wav_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the wav_data.mid file in the project MID directory. Once the 
%%  wav_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the ecg function

\name{ecg}
\alias{ecg}
\title{Electrocardiogram Data}
\concept{measured time series}
\description{This 2048 point ECG data series represents approximately 15 beats
of a normal human cardiac rhythm. This sequence is sampled
at \code{180} Hz and is in units of millivolts. This series was collected
and supplied by Dr. G. Bardy M.D. and Dr. P. Reinhall of the
University of Washington.}

\seealso{
\code{\link{nile}}, \code{\link{atomclock}}, \code{\link{ocean}}, \code{\link{fdp045}}.}

\examples{
plot(ecg)
}
\keyword{ts}

