%%  WARNING: This file was automatically generated from the associated 
%%  wav_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the wav_data.mid file in the project MID directory. Once the 
%%  wav_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the fdp045 function

\name{fdp045}
\alias{fdp045}
\title{Fractional Difference Process Data}
\concept{simulated time series}
\description{These data represent a single realization of a fractional
difference process (FDP) with FDP exponent \eqn{\delta=0.45}{delta=0.45} and a process variance of unity.}

\seealso{
\code{\link{nile}}, \code{\link{atomclock}}, \code{\link{ocean}}, \code{\link{ecg}}.}

\examples{
plot(fdp045)
}
\keyword{ts}

