%%  WARNING: This file was automatically generated from the associated 
%%  wav_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the wav_data.mid file in the project MID directory. Once the 
%%  wav_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the nile function

\name{nile}
\alias{nile}
\title{Yearly Nile river level minima}
\concept{measured time series}
\description{These data represent the measurements of the minimum yearly water
level of the Nile over the years 622 to 1284 as recorded at the Roda gauge near Cairo.}

\seealso{
\code{\link{atomclock}}, \code{\link{ocean}}, \code{\link{ecg}}, \code{\link{fdp045}}.}

\examples{
plot(nile)
}
\keyword{ts}

