%%  WARNING: This file was automatically generated from the associated 
%%  wav_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the wav_data.mid file in the project MID directory. Once the 
%%  wav_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the ocean function

\name{ocean}
\alias{ocean}
\title{Vertical Shear Ocean Data}
\concept{measured time series}
\description{This data is collected by an instrument dropped over the side of a ship
in the ocean. As the instrument descends vertically in the water it
collects information as a function of depth. The ocean data set is a
record of the horizontal (shear) velocity of the water measured in 0.1
increments starting from a depth of 489.5 meters and ending at 899.0
meters. This series was collected and supplied by Mike Gregg, Applied
Physics Laboratory, University of Washington.}

\seealso{
\code{\link{nile}}, \code{\link{atomclock}}, \code{\link{ecg}}, \code{\link{fdp045}}.}

\examples{
plot(ocean)
}
\keyword{ts}

